collapse editions text on editions list
