+++ src/components/Programs/Editions/editionsList.ts
@@ -1,11 +1,11 @@
-import { defineComponent } from "@vue/composition-api"
+import { defineComponent, ref } from "@vue/composition-api"
 import Loader from "@/components/Loader.vue"
 import { MAIN_COLOR } from "@/utils/constants"
 import useEditions from "@/models/useEditions"
 import Edition from "@/components/Programs/Editions/Edition.vue"
 
 export default defineComponent({
-  name: "Editions",
+  name: "EditionsList",
   components: {
     Loader,
     Edition,
