+++ src/components/Programs/ProgramPage/programPage.ts
@@ -5,7 +5,7 @@ import Loader from "@/components/Loader.vue"
 import Media from "@/components/Media/Media.vue"
 import DowneyProgramVideos from "@/components/Programs/DowneyProgramVideos.vue"
 import ProgramActivites from "@/components/Programs/ProgramActivities.vue"
-import Editions from "@/components/Programs/Editions/Editions.vue"
+import Editions from "@/components/Programs/Editions/EditionsList.vue"
 import Residencies from "@/components/Programs/Residencies/Residencies.vue"
 import SchoolProgramsTabs from "@/components/Programs/SchoolProgramsTabs/SchoolProgramsTabs.vue"
 import SchoolProgramWorkshopsTabs from "@/components/Programs/SchoolProgramsTabs/SchoolProgramWorkshopsTabs.vue"
