+++ src/shims-imports.d.ts
@@ -7,5 +7,4 @@ declare module "vue-infinite-loading"
 declare module "vue-ellipsis"
 declare module "vue-burger-button"
 declare module "vue-burger-menu"
-declare module "vue2-collapse"
 declare module "@ivanv/vue-collapse-transition"
